Add particle intersection example. Add onExitRequested hook.
Allow clean up of resources or request that the application not be closed. Update gallery example to not use a tab controller because currently on Nature of Code examples have been written. Update canvas_ops.RawAtlas() to match the optional parameters of the corresponding canvas function. Update mouse handling in a few examples. Use simpler version specifiers in pubspec.yaml.